--- apps/rxjs-web-demo-e2e/.eslintrc
@@ -1,1 +1,0 @@
-{ "extends": "../../.eslintrc", "rules": {} }
