--- apps/rxjs-web-demo-e2e/cypress.json
@@ -1,12 +1,0 @@
-{
-  "fileServerFolder": ".",
-  "fixturesFolder": "./src/fixtures",
-  "integrationFolder": "./src/integration",
-  "modifyObstructiveCode": false,
-  "pluginsFile": "./src/plugins/index",
-  "supportFile": "./src/support/index.ts",
-  "video": true,
-  "videosFolder": "../../dist/cypress/apps/rxjs-web-demo-e2e/videos",
-  "screenshotsFolder": "../../dist/cypress/apps/rxjs-web-demo-e2e/screenshots",
-  "chromeWebSecurity": false
-}
