--- apps/rxjs-web-demo-e2e/src/fixtures/example.json
@@ -1,4 +1,0 @@
-{
-  "name": "Using fixtures to represent data",
-  "email": "hello@cypress.io"
-}
