--- apps/rxjs-web-demo-e2e/tsconfig.e2e.json
@@ -1,8 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "sourceMap": false,
-    "outDir": "../../dist/out-tsc"
-  },
-  "include": ["src/**/*.ts", "src/**/*.js"]
-}
