--- apps/rxjs-web-demo/.eslintrc
@@ -1,1 +1,0 @@
-{ "extends": "../../.eslintrc", "rules": {} }
