--- apps/rxjs-web-demo/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  name: 'rxjs-web-demo',
-  preset: '../../jest.config.js',
-  coverageDirectory: '../../coverage/apps/rxjs-web-demo'
-};
