--- apps/rxjs-web-demo/src/app/app.element.spec.ts
@@ -1,21 +1,0 @@
-import { AppElement } from './app.element';
-
-describe('AppElement', () => {
-  let app: AppElement;
-
-  beforeEach(() => {
-    app = new AppElement();
-  });
-
-  it('should create successfully', () => {
-    expect(app).toBeTruthy();
-  });
-
-  it('should have a greeting', () => {
-    app.connectedCallback();
-
-    expect(app.querySelector('h1').innerHTML).toEqual(
-      'Welcome to rxjs-web-demo!'
-    );
-  });
-});
