--- apps/rxjs-web-demo/src/environments/environment.ts
@@ -1,6 +1,0 @@
-// This file can be replaced during build by using the `fileReplacements` array.
-// When building for production, this file is replaced with `environment.prod.ts`.
-
-export const environment = {
-  production: false
-};
