--- apps/rxjs-web-demo/src/main.ts
@@ -1,1 +1,0 @@
-import './app/app.element.ts';
