--- apps/rxjs-web-demo/src/test-setup.ts
@@ -1,1 +1,0 @@
-import 'document-register-element';
