+++ libs/rxjs-web/package.json
@@ -4,5 +4,13 @@
   "dependencies": {
     "@types/resize-observer-browser": "^0.1.3"
   },
-  "type": "**/*.d.ts"
+  "type": "src/index.ts",
+  "author": "Jan-Niklas Wortmann <6104311+JWO719@users.noreply.github.com>",
+  "repository": {
+    "type": "git",
+    "url": ""
+  },
+  "engines": {
+    "node": ">=10.0.0"
+  }
 }
