+++ libs/rxjs-web/src/lib/types/observer.d.ts
@@ -11,31 +11,3 @@ export interface ObserverNotification<T = any, O = any> {
    */
   observer: Readonly<O>;
 }
-export type IntersectionNotification = ObserverNotification<
-  IntersectionObserverEntry,
-  IntersectionObserver
->;
-
-export type MutationNotification = ObserverNotification<
-  MutationRecord,
-  MutationObserver
->;
-
-export type ResizeNotification = ObserverNotification<
-  ResizeObserverEntry,
-  ResizeObserver
->;
-
-/**
- * Performance Observer notification
- */
-export interface PerformanceNotification {
-  /**
-   * A performance observer entry list
-   */
-  entries: PerformanceObserverEntryList;
-  /**
-   * The Performance Observer
-   */
-  observer: PerformanceObserver;
-}
