+++ nx.json
@@ -13,13 +13,6 @@
   "projects": {
     "rxjs-web": {
       "tags": []
-    },
-    "rxjs-web-demo": {
-      "tags": []
-    },
-    "rxjs-web-demo-e2e": {
-      "tags": [],
-      "implicitDependencies": ["rxjs-web-demo"]
     }
   }
 }
