+++ workspace.json
@@ -38,115 +38,6 @@
           }
         }
       }
-    },
-    "rxjs-web-demo": {
-      "root": "apps/rxjs-web-demo",
-      "sourceRoot": "apps/rxjs-web-demo/src",
-      "projectType": "application",
-      "schematics": {},
-      "architect": {
-        "build": {
-          "builder": "@nrwl/web:build",
-          "options": {
-            "outputPath": "dist/apps/rxjs-web-demo",
-            "index": "apps/rxjs-web-demo/src/index.html",
-            "main": "apps/rxjs-web-demo/src/main.ts",
-            "polyfills": "apps/rxjs-web-demo/src/polyfills.ts",
-            "tsConfig": "apps/rxjs-web-demo/tsconfig.app.json",
-            "assets": [
-              "apps/rxjs-web-demo/src/favicon.ico",
-              "apps/rxjs-web-demo/src/assets"
-            ],
-            "styles": ["apps/rxjs-web-demo/src/styles.scss"],
-            "scripts": []
-          },
-          "configurations": {
-            "production": {
-              "fileReplacements": [
-                {
-                  "replace": "apps/rxjs-web-demo/src/environments/environment.ts",
-                  "with": "apps/rxjs-web-demo/src/environments/environment.prod.ts"
-                }
-              ],
-              "optimization": true,
-              "outputHashing": "all",
-              "sourceMap": false,
-              "extractCss": true,
-              "namedChunks": false,
-              "extractLicenses": true,
-              "vendorChunk": false,
-              "budgets": [
-                {
-                  "type": "initial",
-                  "maximumWarning": "2mb",
-                  "maximumError": "5mb"
-                }
-              ]
-            }
-          }
-        },
-        "serve": {
-          "builder": "@nrwl/web:dev-server",
-          "options": {
-            "buildTarget": "rxjs-web-demo:build"
-          },
-          "configurations": {
-            "production": {
-              "buildTarget": "rxjs-web-demo:build:production"
-            }
-          }
-        },
-        "lint": {
-          "builder": "@nrwl/linter:lint",
-          "options": {
-            "linter": "eslint",
-            "config": "apps/rxjs-web-demo/.eslintrc",
-            "tsConfig": [
-              "apps/rxjs-web-demo/tsconfig.app.json",
-              "apps/rxjs-web-demo/tsconfig.spec.json"
-            ],
-            "exclude": ["**/node_modules/**", "!apps/rxjs-web-demo/**"]
-          }
-        },
-        "test": {
-          "builder": "@nrwl/jest:jest",
-          "options": {
-            "jestConfig": "apps/rxjs-web-demo/jest.config.js",
-            "tsConfig": "apps/rxjs-web-demo/tsconfig.spec.json",
-            "passWithNoTests": true,
-            "setupFile": "apps/rxjs-web-demo/src/test-setup.ts"
-          }
-        }
-      }
-    },
-    "rxjs-web-demo-e2e": {
-      "root": "apps/rxjs-web-demo-e2e",
-      "sourceRoot": "apps/rxjs-web-demo-e2e/src",
-      "projectType": "application",
-      "architect": {
-        "e2e": {
-          "builder": "@nrwl/cypress:cypress",
-          "options": {
-            "cypressConfig": "apps/rxjs-web-demo-e2e/cypress.json",
-            "tsConfig": "apps/rxjs-web-demo-e2e/tsconfig.e2e.json",
-            "devServerTarget": "rxjs-web-demo:serve"
-          },
-          "configurations": {
-            "production": {
-              "devServerTarget": "rxjs-web-demo:serve:production"
-            }
-          }
-        },
-        "lint": {
-          "builder": "@nrwl/linter:lint",
-          "options": {
-            "linter": "eslint",
-            "config": "apps/rxjs-web-demo-e2e/.eslintrc",
-            "tsConfig": ["apps/rxjs-web-demo-e2e/tsconfig.e2e.json"],
-            "exclude": ["**/node_modules/**", "!apps/rxjs-web-demo-e2e/**"]
-          }
-        }
-      }
     }
   },
   "cli": {
